integrations/paddle-dodo/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pydantic
pydantic-settings
python-dotenv
PyJWT
email-validator